refactor profiles db table, update createOrder, fix horder history, update copied traders tab layout, add comments

+++ internal/user/copied-traders.go
@@ -60,7 +60,7 @@ type Order struct {
 	Qty            float64 `json:"qty"`
 	Price          float64 `json:"price"`
 	Side           string  `json:"side"`
-	IsLeverage     float64 `json:"isLeverage"`
+	IsLeverage     int64   `json:"isLeverage"`
 }
 
 type CopiedTradersManager struct {
